requirements.txt
bleach
markdown
django-simple-captcha
django-contrib-comments
django-mptt
